Replace a number of stdint variables with standard C types.
uint8_t sticks around for raw data, and u/int64_t is for big values,
but most everything else works as ints/unsigneds.

+++ src/common/flac_crc.h
@@ -4,7 +4,7 @@
   assigns a new checksum to that value*/
 
 void
-flac_crc8(uint8_t byte, unsigned *checksum);
+flac_crc8(uint8_t byte, uint8_t *checksum);
 
 void
-flac_crc16(uint8_t byte, unsigned *checksum);
+flac_crc16(uint8_t byte, uint16_t *checksum);
